requirements.txt
python-telegram-bot==20.6
python-dotenv
aiohttp
spotdl==4.2.10